circular dependancy with structs
